Update DamCatalogService and related tests
Additional annotation imports in DamCatalogRestController have been added. DamCatalogService class and its methods are re-annotated for a better description clarity. Logger imports which were not used in DailyMeasurementService are removed. A new test class DamCatalogServiceTest is created which includes unit tests for the DamCatalogService. Minor corrections are made in DamInfoServiceTest.